+++ app/containers/PhoneControlsInactive/tests/index.test.js
@@ -1,0 +1,11 @@
+// import { PhoneControlsInactive } from '../index';
+
+import expect from 'expect';
+// import { shallow } from 'enzyme';
+// import React from 'react';
+
+describe('<PhoneControlsInactive />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(false);
+  });
+});